Update bank_system .sql
id type changed from int to bigint

+++ bank_system .sql
@@ -191,7 +191,7 @@ END;
 $show_payment_log$ LANGUAGE plpgsql
 
 --procedure
-CREATE OR REPLACE PROCEDURE withdrawAmount(id int,amt NUMERIC(12,3))
+CREATE OR REPLACE PROCEDURE withdrawAmount(id BIGINT,amt NUMERIC(12,3))
 language plpgsql
 as $$
 declare
